requirements.txt
Flask
Flask-SQLAlchemy
Flask-SocketIO>=5.3.0
python-socketio>=5.7.0
eventlet>=0.33.0
Werkzeug
Pillow
qrcode
python-dotenv
itsdangerous
requests
Flask-Migrate
psycopg2-binary
Flask-Caching
pywebpush
Flask-WTF
Flask-Limiter
fpdf2
pandas
